fix: release the root command
