+++ CONTRIBUTING.md
@@ -13,7 +13,6 @@
 
 - [Golang](https://golang.org/)
 - [Cobra](https://github.com/spf13/cobra)
-- [Viper](https://github.com/spf13/viper)
 
 ## 💥 How to Contribute
